requirements.txt
openpyxl==3.1.5
pandas==2.2.3
python-decouple==3.8
rapidfuzz==3.11.0
Requests==2.32.3
streamlit==1.41.1
Unidecode==1.3.8
